please fix

+++ include/sys/filedirectory.h
@@ -8,8 +8,8 @@ char *filename;
 struct filedir* parent;
 struct filedir* root;
 struct filedir* children[10];
-uint64_t type; //1: directory , 2:file
-uint64_t child_count; 
+int type; //1: directory , 2:file
+int child_count; 
 }__attribute__((packed))filedir;
 
 filedir *fileDescriptor;
